backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-jwt-extended==4.6.0
bcrypt==4.1.3
google-generativeai==0.7.2
python-dotenv==1.0.1
pymongo==4.8.0
